ddi_encode_vp9: little hack to deal with multi type buffer segment list
VAEncCodedBufferSegment with multiple types of buffers is unhandled
in many of the software projects. For eg: many of the middleware
implementations check the size of VACodedBufferSegment and if it is
non-zero, the new buffer will get interpreted as encoded buffer rather
than status buffer which leads to seg fault. So this is a temporary
workaround to make the customers happy.

codedBufferSegment->next = NULL;

Actual deallocation will happen later during DdiEncodeVp9 destruction.